move shorthand table resolver to /shorthand, make full table resolver more standalone
