+++ packages/store/ts/config/v2/index.ts
@@ -1,9 +1,9 @@
 export * from "./generics";
 export * from "./scope";
 export * from "./schema";
-export * from "./tableShorthand";
-export * from "./tableFull";
+export * from "./shorthand/tableShorthand";
 export * from "./table";
+export * from "./shorthand/table";
 export * from "./store";
 export * from "./output";
 export * from "./defaults";
