+++ packages/store/ts/config/v2/shorthand/tableShorthand.test.ts
@@ -1,6 +1,6 @@
 import { describe, it } from "vitest";
 import { attest } from "@arktype/attest";
-import { AbiTypeScope, extendScope } from "./scope";
+import { AbiTypeScope, extendScope } from "../scope";
 import { resolveTableShorthand } from "./tableShorthand";
 
 describe("resolveTableShorthand", () => {
