--- packages/store/ts/config/v2/tableFull.test.ts
@@ -1,43 +1,0 @@
-import { attest } from "@arktype/attest";
-import { describe, it } from "vitest";
-import { getStaticAbiTypeKeys, AbiTypeScope, extendScope } from "./scope";
-import { validateKeys } from "./tableFull";
-
-describe("validateKeys", () => {
-  it("should return a tuple of valid keys", () => {
-    attest<
-      ["static"],
-      validateKeys<getStaticAbiTypeKeys<{ static: "uint256"; dynamic: "string" }, AbiTypeScope>, ["static"]>
-    >();
-  });
-
-  it("should return a tuple of valid keys with an extended scope", () => {
-    const scope = extendScope(AbiTypeScope, { static: "address", dynamic: "string" });
-
-    attest<
-      ["static", "customStatic"],
-      validateKeys<
-        getStaticAbiTypeKeys<
-          { static: "uint256"; dynamic: "string"; customStatic: "static"; customDynamic: "dynamic" },
-          typeof scope
-        >,
-        ["static", "customStatic"]
-      >
-    >();
-  });
-
-  it("should return a tuple of valid keys with an extended scope", () => {
-    const scope = extendScope(AbiTypeScope, { static: "address", dynamic: "string" });
-
-    attest<
-      ["static", "customStatic"],
-      validateKeys<
-        getStaticAbiTypeKeys<
-          { static: "uint256"; dynamic: "string"; customStatic: "static"; customDynamic: "dynamic" },
-          typeof scope
-        >,
-        ["static", "customStatic"]
-      >
-    >();
-  });
-});
